Restructured the UiMenu Code into several files Simplified some small code pieces but basically there is now a separation between menu handling (ui_menu_internal.c), menu structure definition (ui_menu_structure.c), and item change logic (ui_menu.c).
The external interface to the menu system is clearly separated from the
internal data structures and functions.

Tested briefly, looks good and compiles on all IDEs
