+++ mchf-eclipse/drivers/ui/radio_management.h
@@ -86,14 +86,13 @@ typedef enum
 typedef struct BandInfo
 {
     uint8_t default_pf;
-    uint8_t selector_5W;
-    uint8_t selector_FULL;
     uint32_t tune;
     uint32_t size;
     const char* name;
 } BandInfo;
 
 extern const BandInfo bandInfo[MAX_BAND_NUM];
+// FIXME: for technical reasons defined in ui_menu.c
 
 
 // SWR and RF power meter public
